+++ src/rom/RomOnly.cpp
@@ -41,10 +41,10 @@ void RomOnly::write_u8(uint16_t addr, uint8_t val) {
 }
 
 void RomOnly::serialize(std::ofstream &f) {
-	for (int i = 0; i < rom_banks.size(); i++) {
+	for (size_t i = 0; i < rom_banks.size(); i++) {
 		f.write(reinterpret_cast<const char *>(&rom_banks[i]), sizeof(rom_banks[i]));
 	}
-	for (int i = 0; i < ram_banks.size(); i++) {
+	for (size_t i = 0; i < ram_banks.size(); i++) {
 		f.write(reinterpret_cast<const char *>(&ram_banks[i]), sizeof(ram_banks[i]));
 	}
 	f.write(reinterpret_cast<const char *>(&ram_enable), sizeof(ram_enable));
@@ -52,10 +52,10 @@ void RomOnly::serialize(std::ofstream &f) {
 }
 
 void RomOnly::deserialize(std::ifstream &f) {
-	for (int i = 0; i < rom_banks.size(); i++) {
+	for (size_t i = 0; i < rom_banks.size(); i++) {
 		f.read(reinterpret_cast<char *>(&rom_banks[i]), sizeof(rom_banks[i]));
 	}
-	for (int i = 0; i < ram_banks.size(); i++) {
+	for (size_t i = 0; i < ram_banks.size(); i++) {
 		f.read(reinterpret_cast<char *>(&ram_banks[i]), sizeof(ram_banks[i]));
 	}
 	f.read(reinterpret_cast<char *>(&ram_enable), sizeof(ram_enable));
